Further work on the calories calculators
- Save the user's data to the profile when performing the calculations
- The different calculations are done by the model
- There is a very rough indication of planed calories on the plan view

+++ wger/nutrition/forms.py
@@ -53,7 +53,7 @@ class UnitChooserForm(forms.Form):
 
 
 class BmiForm(forms.ModelForm):
-    weight = forms.IntegerField()
+    weight = forms.DecimalField()
 
     class Meta:
         model = UserProfile
@@ -64,7 +64,7 @@ class BmrForm(forms.ModelForm):
     '''
     Form for the basal metabolic rate
     '''
-    weight = forms.IntegerField()
+    weight = forms.DecimalField()
 
     class Meta:
         model = UserProfile
